requirements.txt
Flask
argon2-cffi
pycryptodome
riposte